+++ exolve-to-ipuz.js
@@ -24,7 +24,7 @@ SOFTWARE.
 The latest code and documentation for Exolve can be found at:
 https://github.com/viresh-ratnakar/exolve
 
-Version: Exolve v1.64, November 25, 2025
+Version: Exolve v1.65, November 25, 2025
 */
 
 function exolveToIpuzAddClue(xlvp, ci, cluesArray) {